Create <CodeBlock /> component
